Group *_member_group.dart into one directory
