+++ lib/lints/member_ordering/models/member_group/method_member_group.dart
@@ -24,7 +24,7 @@
 import 'package:analyzer/dart/ast/ast.dart' show Identifier, MethodDeclaration;
 import 'package:solid_lints/lints/member_ordering/member_ordering_utils.dart';
 import 'package:solid_lints/lints/member_ordering/models/annotation.dart';
-import 'package:solid_lints/lints/member_ordering/models/member_group.dart';
+import 'package:solid_lints/lints/member_ordering/models/member_group/member_group.dart';
 import 'package:solid_lints/lints/member_ordering/models/member_type.dart';
 import 'package:solid_lints/lints/member_ordering/models/modifier.dart';
